Add index page for client

+++ evaluate_api/app.py
@@ -7,7 +7,6 @@ from evaluate_api.controller import challenges
 from flasgger import Swagger
 from evaluate_api.swagger.swagger import swagger_config
 from evaluate_api.database import engine, db_session, Base
-import evaluate_api.model
 
 """
 for convert structure to api style
@@ -18,12 +17,15 @@ check this page: https://flask.palletsprojects.com/en/2.1.x/views/#method-views-
 def create_app(test_config=None):
     """
     entry for flask application
+    if reset_database == True => drop tables and regenerate from files
 
     Args:
         test_config (dict): config for test if none set as default
     Returns
         app (Flask): generated flask application
     """
+    reset_database = False
+
     app = Flask(__name__, instance_relative_config=True, template_folder=r"./templates")
 
     if test_config is None:
@@ -37,10 +39,11 @@ def create_app(test_config=None):
     else:
         app.config.from_mapping(test_config)
 
-    Base.metadata.drop_all(bind=engine)
-    Base.metadata.create_all(bind=engine)
-    load_challenges_to_db(db_session)
-    load_results_to_db(db_session)
+    if reset_database:
+        Base.metadata.drop_all(bind=engine)
+        Base.metadata.create_all(bind=engine)
+        load_challenges_to_db(db_session)
+        load_results_to_db(db_session)
 
     app.register_blueprint(challenges)
     swagger = Swagger(app, config=swagger_config, template_file="./swagger/swagger.yaml")
@@ -58,6 +61,10 @@ def create_app(test_config=None):
     def handle_500(e):
         print(e)
         return jsonify({'error': 'Something went wrong, we are working on it'}), 500
+
+    @app.teardown_request
+    def shutdown_session(exception=None):
+        db_session.remove()
 
     return app
